Tried to botch together some simple visuals with what is currently in the system. It half-way works. In the process of doing so, I found some bugs but also made the whole system messier. TODO for next commit: Fix the vec3 vec4 discrepency between host and device and make it nice, go through everything and start to add proper support for some of the features you haven't put in yet (entity types, materials). Also, figure out a system for making the Entity struct cleaner, considering we're going to be filling it up when we add entity types (use unions, those are good). Cache some of the things that you don't have to calculate on the GPU, like halfWindowWidth. Delete most of what you have in your kernel and start fresh, make a proper system that handles every different material through a main material switch case somewhere. When you have a good, easy to use system set up, then you can start worrying about adding kd-tree. ALSO, add antialiasing support and dynamic sampling support.

+++ fractal/Shader.h
@@ -30,8 +30,8 @@ public:
 	virtual void setCameraRotation(nmath::Vector3f rotation) = 0;
 	virtual void setCameraFOV(float FOV) = 0;
 
-	virtual void setScene(cl_mem computeScene, size_t computeSceneLength) = 0;
+	virtual void setScene(cl_mem computeScene, uint64_t computeSceneLength) = 0;
 
-	virtual void setMaterialHeap(cl_mem computeMaterialHeap, size_t computeMaterialHeapLength) = 0;
-	virtual void setMaterialHeapOffset(size_t computeMaterialHeapOffset) = 0;
+	virtual void setMaterialHeap(cl_mem computeMaterialHeap, uint64_t computeMaterialHeapLength) = 0;
+	virtual void setMaterialHeapOffset(uint64_t computeMaterialHeapOffset) = 0;
 };
